+++ src/main/java/lotto/domain/numberlotto/constants/message/Message.java
@@ -1,4 +1,4 @@
-package lotto.domain.constants.message;
+package lotto.domain.numberlotto.constants.message;
 
 public interface Message {
