+++ src/test/java/lotto/LottoTest.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.domain.lotto.Lotto;
+import lotto.domain.numberlotto.lotto.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
